# Refactor/Clean-Up Composer and Improve LEAN Startup Performance