… Common/Util/Composer.cs