<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
This is mostly a clean up/refactoring of the existing logic and behavior.
The only real behavior change is not waiting for the CompositionContainer
to be fully resolved before finishing the init procedures. 90% of the call
we make to the Composer are QC types and therefore handled by the special
case code that explicitly loads QC assemblies and caches their types.

Another behavior change was made to really trim down the number of types
that were being cached from the QC assemblies. We were caching ALL public
types that weren't abstract/interface/enum, but since it's a proxy for MEF
we should really have the [InheritedExport] filter added. Adding this cut
down the number of QC types were caching by over 100x. Another perf gain
was in how we store the cached items. The existing code had the cache in
a list with over 10k types in it and then would search that list doing
an IsAssignableFrom and MatchesTypeName. That's a lot of items to hit.
This commit changes the storage to be a dictionarty keyed by the contract
type. We use the type that declared the [InheritedExport] as well as the
implementation type to support requested exact types, but I'm not sure
that this was required but at first glance it seemed to be supported by
the existing implementation.

This change cuts down LEAN's startup time (defined as the time it takes
to go from the top of Main to just following the initialization/resolution
of the algorithm handlers. This section of code's performance is 100%
dominated by the Composer's initialization and resolution functions. Avg
times on 'master' were around 350ms for a hot start and over 1.5s for a
cold start. Cold start meaning first run after fresh build. This change
shows an average hot start time of about 210ms and a cold start time of
about 350ms

Closes #3836

I've also included a `Composer.HardReset()` method that resets all state within the composer, including reloading assemblies from disk.

Surprisingly, using:
``` csharp
public static readonly Composer Instance = new Composer();
``` 
instead of a `Lazy<Composer>` or even _just_ a computed property, saved over ~150ms!! I never would have expected that to provide such a larger performance improvement.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
See #3836 

#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
While I'm not a fan of the `Composer` these days and would love to replace it... probably with a simple type name lookup to remove MEF... I still hold some nostalgia towards it, so it felt good to clean it up and make it look a little prettier. I initially went to look at it because it was taking ~15 seconds to start LEAN, but that was for a completely unrelated reason.

#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
No.

#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Tested using the existing unit tests as well as thorough performance testing, including profiling as well as timing LEAN startup. See the issue for the timing snippet used.

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] Refactor w/ added bonus of performance improvement 😛 
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->