@mchandschuh 2019-11-15
Taking this down to focus on the WFO PR. AFAIK, there wasn't a real issue, I just noticed some things. For posterity -- there was an error in the compiler with these changes:
> System.IndexOutOfRangeException: Index was outside the bounds of the array.
Also, there seems to be an issue with running the entire test suit w/ it hanging on launch -- likely a side-effect of an error being thrown from within a static initializer (all static field initializers are essentially executed via static constructors)